@@ -1,12 +1,12 @@
 <?xml version="1.0"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<groupId>com.santiagolizardo.madcommander</groupId>
-	<artifactId>MadCommander</artifactId>
+	<groupId>com.santiagolizardo.neocommander</groupId>
+	<artifactId>NeoCommander</artifactId>
 	<version>1.4.7</version>
 	<packaging>jar</packaging>
-	<name>MadCommander</name>
-	<url>http://sourceforge.net/projects/madcommander/</url>
+	<name>NeoCommander</name>
+	<url>https://github.com/santiagolizardo/neocommander</url>
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 	</properties>
@@ -35,8 +35,8 @@
 				<artifactId>maven-compiler-plugin</artifactId>
 				<version>3.1</version>
 				<configuration>
-					<source>1.7</source>
-					<target>1.7</target>
+					<source>1.8</source>
+					<target>1.8</target>
 				</configuration>
 			</plugin>
 			<plugin>
