@@ -32,10 +32,8 @@ import com.santiagolizardo.madcommander.resources.images.IconFactory;
 
 public class FLCellRenderer extends DefaultTableCellRenderer {
 
-	/**
-	 * 
-	 */
 	private static final long serialVersionUID = 8224673036449358242L;
+
 	private Format format;
 
 	public FLCellRenderer(Format format) {
