@@ -16,24 +16,23 @@
  */
 package com.santiagolizardo.madcommander.util;
 
-import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
-import junit.framework.Assert;
+import org.junit.Assert;
+import org.junit.Test;
 
-/**
- *
- * @author santiago
- */
 public class ListUtilsTest {
 
+	@Test
 	public void testImplode() {
 		List<String> names = Arrays.asList(new String[]{"Foo", "Bar", "John", "Doe"});
-		Assert.assertEquals("Foo,Bar,John,Doe", ListsUtils.implode(",", names));
-		Assert.assertEquals("FooBarJohnDoe", ListsUtils.implode("", names));
-		Assert.assertEquals("", ListsUtils.implode(",", new ArrayList()));
+		Assert.assertEquals("Foo,Bar,John,Doe", String.join(",", names));
+		Assert.assertEquals("FooBarJohnDoe", String.join("", names));
+		Assert.assertEquals("", String.join(",", Collections.emptyList()));
 	}
 
+	@Test
 	public void testExplode() {
 		List<String> names = ListsUtils.explode(",", "Foo,Bar,John,Doe");
 		Assert.assertEquals(4, names.size());
